refactor(dashboard): split state.ts and shrink createDashboardClient
state.ts becomes the discriminated reducer (329 lines); SDK→view-model
mappers move to mappers.ts (191 lines); view-model derivation plumbing
moves to view-model.ts (217 lines). All public exports preserved.

createDashboardClient drops from 83 to ~44 lines by extracting
fetchMetricsFanOut and a listUrl helper.

Both keep the complexity check within the existing ratchet budget.
